Revert change from 07b10ec
Clearly this was important to the overall function of this strategy but I'm not really sure I understand why

In the above commit surely we would want to check for a steady decline over STRIDE values of stoch K+D but the opposite is true?

Needs more research. Do not touch :^)